Move VersionsTable
