+++ site/src/pages/TemplatePage/TemplateVersionsPage/TemplateVersionsPage.tsx
@@ -4,7 +4,7 @@ import { getErrorMessage } from "api/errors"
 import { ConfirmDialog } from "components/Dialogs/ConfirmDialog/ConfirmDialog"
 import { displayError, displaySuccess } from "components/GlobalSnackbar/utils"
 import { useTemplateLayoutContext } from "components/TemplateLayout/TemplateLayout"
-import { VersionsTable } from "components/VersionsTable/VersionsTable"
+import { VersionsTable } from "./VersionsTable"
 import { useState } from "react"
 import { Helmet } from "react-helmet-async"
 import { getTemplatePageTitle } from "../utils"
